+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # GKI Salatiga+ Changelog
 
+## Next Alpha Update
+
+- New: Added new screen: agenda
+- Improved: Added representative icons to main menu items
+
 ## v0.2.0-alpha (17) --- 2024-08-03
 
 - New: The main menu now has an elegant scrolling behavior